fix(precision): eradicate decimal truncation and enforce IEEE 754 full mantissa
TK-01. Zero `round()` / tensor-round in backend/app; every float that reaches
logs, telemetry, JSON or the API is serialized with full mantissa (.17g).

- firewall: scalar cosine now accumulates in float64 (float32 collapsed the
  1e-5 coordinate gaps to exactly 1.0) and no longer clips, so separation is
  preserved. Zero-norm/alpha logs emit .17g.
- chat/audit: q/c vectors upcast to float64, RAG top-hit score computed in
  float64, telemetry + tuning hints + BurstDetectionBreach use full mantissa.
- recommended_thresholds: sweep grids keep exact np.arange samples instead of
  round(x, 1|2); calibration drops round(accuracy|f1|youden, 4).
- dataset_generator: leakage cosine in float64.
- tests: add test_numerical_purity.py (18 checks: AST round/format/float16
  guards, micro-gap survival, grid exactness, helper dtype) and update the
  legacy tuning-hint assertion that hard-coded the truncated rendering.

+++ backend/app/api/endpoints/config.py
@@ -143,8 +143,8 @@ async def audit_query(request: Request, req: AuditRequest):
     context = join_rag_context(context_chunks)
 
     c_vec = results[0]["vector"]
-    q_arr = np.array(q_vec, dtype=np.float32)
-    c_arr = np.array(c_vec, dtype=np.float32)
+    q_arr = np.array(q_vec, dtype=np.float64)
+    c_arr = np.array(c_vec, dtype=np.float64)
     word_count = len(req.query.split())
 
     result = SemanticFirewall.evaluate_clause(q_arr, c_arr, cfg, word_count)
